GetTwitchUser/requirements.txt
Flask==2.0.3
requests==2.26.0
Werkzeug==2.0.3
psycopg2-binary==2.9.9
